+++ Backend/src/main/java/edu/depaul/cdm/se452/rfa/authentication/entity/UserRole.java
@@ -14,7 +14,6 @@ public class UserRole {
     private UserRoleId id;
 
     @MapsId("userId")
-    @JsonIgnore
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "user_id", nullable = false)
     @JsonIgnore
