+++ Backend/src/main/java/edu/depaul/cdm/se452/rfa/profileManagement/entity/Profile.java
@@ -23,7 +23,7 @@ public class Profile {
     private Integer id;
 
     @NotNull
-    @OneToOne(fetch = FetchType.LAZY, optional = false)
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
     @OnDelete(action = OnDeleteAction.CASCADE)
     @JoinColumn(name = "user_id", nullable = false)
     private User user;
